bento_service/requirements.txt
bentoml==1.2.18
pandas==2.2.3
scikit-learn==1.3.2
pydantic==1.10.15
python-dotenv==1.0.1
boto3==1.35.15
numpy==1.26.4
scipy==1.11.4
